Prim’s_Algorithm_to_Find_the_Minimum_Spanning_Tree_of_a_Graph.cpp
This is a C++ Program to find the minimum spanning tree of the given graph using Prim's algorihtm. In computer science, Prim’s algorithm is a greedy algorithm that finds a minimum spanning tree for a connected weighted undirected graph. This means it finds a subset of the edges that forms a tree that includes every vertex, where the total weight of all the edges in the tree is minimized.